Move `TextSelectCopy` to design package
