+++ web/packages/design/src/TextSelectCopy/TextSelectCopy.tsx
@@ -16,11 +16,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React from 'react';
+import { useRef, useState } from 'react';
+import { useTheme } from 'styled-components';
+
 import { copyToClipboard } from 'design/utils/copyToClipboard';
 import selectElementContent from 'design/utils/selectElementContent';
 import { ButtonPrimary, Box, Flex } from 'design';
-import { useTheme } from 'styled-components';
 
 export function TextSelectCopy({
   text,
@@ -31,8 +32,8 @@ export function TextSelectCopy({
   ...styles
 }: Props) {
   const font = fontFamily || useTheme().fonts.mono;
-  const ref = React.useRef();
-  const [copyCmd, setCopyCmd] = React.useState(() => 'Copy');
+  const ref = useRef();
+  const [copyCmd, setCopyCmd] = useState(() => 'Copy');
 
   function onCopyClick() {
     copyToClipboard(text).then(() => setCopyCmd('Copied'));
