+++ web/packages/design/src/TextSelectCopy/TextSelectCopyMulti.story.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 import { TextSelectCopyMulti as Component } from './TextSelectCopyMulti';
 
 export default {
-  title: 'Teleport/TextSelectCopy/Multi',
+  title: 'Design/TextSelectCopy/Multi',
 };
 
 export const BashMulti = () => {
